+++ src/Awaitable/Internal/FulfilledAwaitable.php
@@ -9,9 +9,9 @@
 
 namespace Icicle\Awaitable\Internal;
 
-use Icicle\Loop;
-use Icicle\Awaitable\{Awaitable, Promise};
+use Icicle\Awaitable\{Awaitable, Promise, function resolve, function reject};
 use Icicle\Exception\InvalidArgumentError;
+use Icicle\Loop;
 use Throwable;
 
 class FulfilledAwaitable extends ResolvedAwaitable
@@ -43,16 +43,12 @@ class FulfilledAwaitable extends ResolvedAwaitable
         if (null === $onFulfilled) {
             return $this;
         }
-        
-        return new Promise(function (callable $resolve, callable $reject) use ($onFulfilled) {
-            Loop\queue(function () use ($resolve, $reject, $onFulfilled) {
-                try {
-                    $resolve($onFulfilled($this->value));
-                } catch (Throwable $exception) {
-                    $reject($exception);
-                }
-            });
-        });
+
+        try {
+            return resolve($onFulfilled($this->value));
+        } catch (Throwable $exception) {
+            return reject($exception);
+        }
     }
     
     /**
@@ -61,7 +57,13 @@ class FulfilledAwaitable extends ResolvedAwaitable
     public function done(callable $onFulfilled = null, callable $onRejected = null)
     {
         if (null !== $onFulfilled) {
-            Loop\queue($onFulfilled, $this->value);
+            try {
+                $onFulfilled($this->value);
+            } catch (Throwable $exception) {
+                Loop\queue(function () use ($exception) {
+                    throw $exception;
+                });
+            }
         }
     }