+++ src/Awaitable/Internal/RejectedAwaitable.php
@@ -9,8 +9,8 @@
 
 namespace Icicle\Awaitable\Internal;
 
+use Icicle\Awaitable\{Awaitable, function resolve, function reject};
 use Icicle\Loop;
-use Icicle\Awaitable\{Awaitable, Exception\RejectedException, Promise};
 use Throwable;
 
 class RejectedAwaitable extends ResolvedAwaitable
@@ -36,16 +36,12 @@ class RejectedAwaitable extends ResolvedAwaitable
         if (null === $onRejected) {
             return $this;
         }
-        
-        return new Promise(function (callable $resolve, callable $reject) use ($onRejected) {
-            Loop\queue(function () use ($resolve, $reject, $onRejected) {
-                try {
-                    $resolve($onRejected($this->exception));
-                } catch (Throwable $exception) {
-                    $reject($exception);
-                }
-            });
-        });
+
+        try {
+            return resolve($onRejected($this->exception));
+        } catch (Throwable $exception) {
+            return reject($exception);
+        }
     }
     
     /**
@@ -54,7 +50,13 @@ class RejectedAwaitable extends ResolvedAwaitable
     public function done(callable $onFulfilled = null, callable $onRejected = null)
     {
         if (null !== $onRejected) {
-            Loop\queue($onRejected, $this->exception);
+            try {
+                $onRejected($this->exception);
+            } catch (Throwable $exception) {
+                Loop\queue(function () use ($exception) {
+                    throw $exception;
+                });
+            }
         } else {
             Loop\queue(function () {
                 throw $this->exception; // Rethrow exception in uncatchable way.