+++ src/Coroutine/Coroutine.php
@@ -15,31 +15,26 @@ use Icicle\Loop;
 use Throwable;
 
 /**
- * This class implements cooperative coroutines using Generators. Coroutines should yield promises to pause execution
- * of the coroutine until the promise has resolved. If the promise is fulfilled, the fulfillment value is sent to the
- * generator. If the promise is rejected, the rejection exception is thrown into the generator.
+ * This class implements cooperative coroutines using Generators. Coroutines should yield awaitables to pause execution
+ * of the coroutine until the awaitable has resolved. If the awaitable is fulfilled, the fulfillment value is sent to
+ * the generator. If the awaitable is rejected, the rejection exception is thrown into the generator.
  */
 final class Coroutine extends Future
 {
     /**
-     * @var \Generator|null
+     * @var \Generator
      */
     private $generator;
 
     /**
-     * @var \Closure|null
+     * @var \Closure
      */
     private $send;
     
     /**
-     * @var \Closure|null
+     * @var \Closure
      */
     private $capture;
-    
-    /**
-     * @var mixed[]|null
-     */
-    private $next;
 
     /**
      * @var bool
@@ -59,8 +54,8 @@ final class Coroutine extends Future
          * @param mixed $value The value to send to the generator.
          */
         $this->send = function ($value = null) {
-            if ($this->paused) { // If paused, save callable and value for resuming.
-                $this->next = [$this->send, $value];
+            if ($this->paused) { // Avoid blowing up the call stack by queuing the continuation.
+                Loop\queue($this->send, $value);
                 return;
             }
 
@@ -69,16 +64,15 @@ final class Coroutine extends Future
                 $this->next($this->generator->send($value));
             } catch (Throwable $exception) {
                 $this->reject($exception);
-                $this->close();
             }
         };
 
         /**
-         * @param \Exception $exception Exception to be thrown into the generator.
+         * @param \Throwable $exception Exception to be thrown into the generator.
          */
         $this->capture = function (Throwable $exception) {
-            if ($this->paused) { // If paused, save callable and exception for resuming.
-                $this->next = [$this->capture, $exception];
+            if ($this->paused) { // Avoid blowing up the call stack by queuing the continuation.
+                Loop\queue($this->capture, $exception);
                 return;
             }
 
@@ -87,7 +81,6 @@ final class Coroutine extends Future
                 $this->next($this->generator->throw($exception));
             } catch (Throwable $exception) {
                 $this->reject($exception);
-                $this->close();
             }
         };
 
@@ -95,7 +88,6 @@ final class Coroutine extends Future
             $this->next($this->generator->current());
         } catch (Throwable $exception) {
             $this->reject($exception);
-            $this->close();
         }
     }
 
@@ -108,9 +100,10 @@ final class Coroutine extends Future
     {
         if (!$this->generator->valid()) {
             $this->resolve($this->generator->getReturn());
-            $this->close();
             return;
         }
+
+        $this->paused = true;
 
         if ($yielded instanceof Generator) {
             $yielded = new self($yielded);
@@ -121,52 +114,8 @@ final class Coroutine extends Future
         } else {
             Loop\queue($this->send, $yielded);
         }
-    }
 
-    /**
-     * The garbage collector does not automatically detect (at least not quickly) the circular references that can be
-     * created, so explicitly setting these parameters to null is necessary for proper freeing of memory.
-     */
-    private function close()
-    {
-        $this->next = null;
-        $this->send = null;
-        $this->capture = null;
-
-        $this->paused = true;
-
-        $this->generator = null;
-    }
-
-    /**
-     * {@inheritdoc}
-     */
-    public function pause()
-    {
-        $this->paused = true;
-    }
-    
-    /**
-     * {@inheritdoc}
-     */
-    public function resume()
-    {
-        if ($this->isPending() && $this->paused) {
-            $this->paused = false;
-            
-            if (null !== $this->next) {
-                Loop\queue(...$this->next);
-                $this->next = null;
-            }
-        }
-    }
-    
-    /**
-     * {@inheritdoc}
-     */
-    public function isPaused(): bool
-    {
-        return $this->paused;
+        $this->paused = false;
     }
 
     /**
@@ -174,16 +123,10 @@ final class Coroutine extends Future
      */
     public function cancel(Throwable $reason = null)
     {
-        if (null !== $this->generator) {
+        if ($this->isPending()) {
             $current = $this->generator->current(); // Get last yielded value.
             if ($current instanceof Awaitable) {
                 $current->cancel($reason); // Cancel last yielded awaitable.
-            }
-
-            try {
-                $this->close(); // Throwing finally blocks in the Generator may cause close() to throw.
-            } catch (Throwable $exception) {
-                $reason = $exception;
             }
         }